[tests] Remove `XASdkDeployTests`
Most of our time intensive tests are in deploy and debug scenarios, and
we have some duplication there that can be cleaned up.

The `XASdkDeployTests` class has been removed, and non duplicated tests
from this class have been moved into `InstallAndRunTests`.

The `DotNetInstallAndRun` tests have been paired down to two tests for
`net7.0-android` Debug and Release.  The other permutations should be
covered by existing install and run tests.

The `DotNetDebug` tests have been removed, as they were effectively
duplicates of `ApplicationRunsWithDebuggerAndBreaks`.

The `Debugger` tests have had their connection settings adjusted to try
fewer times to connect but wait longer in between attempts.  We will now
only wait for up to 5 minutes to connect rather than ~17 minutes.
